Add a default Pull Request template
The templates added in fa8de7104a123ee8392e61d1cacbac99d891345f are
only visible when added as a query parameter to the PR URL, which
means you can forget it. This template should be visible by
default, although it does require you to manually remove the
checklist that's not applicable.