Add VSCode configuration to enable test debugging
Even though this is an IDE-specific file, it contains project-
specific configuration rather than personal preferences, which is
why I committed it. However, since not everybody might be using
VSCode, and since our CI infrastructure does not verify whether it
breaks, feel free to remove it in the future if it is in the way or
no longer works.