Enable TypeScript-assisted linting
This gives ESLint access to type information by running the
TypeScript compiler, allowing for more helpful linting rules.

Do note that all the linted files also had to be included in the
TypeScript configuration. This means that from now on, test files
are also compiled. I could also have chosen to set up a
separate tsconfig.eslint.json, but the overhead isn't currently
that big, so I chose to avoid the clutter of the extra file
instead.

See
https://github.com/typescript-eslint/typescript-eslint/blob/master/docs/getting-started/linting/TYPED_LINTING.md