Add fetchLitDatasetWithAcl
This is an experimental function that fetches both the Resource,
its ACL file, and the applicable fallback ACL file of one of its
Containers, if accessible to the current user.

This is a pretty heavy operation right now that can result in many
HTTP requests. The reason these are all bundled in a single API
call now (at least externally), is that there are spec proposals
under discussion that would make the server return all of these in
a single call, in which case we could keep this API the same.

If that does not get accepted, we can always expose the underlying
individual ACL fetching functions later, without breaking backwards
compatibility.

(That said, given that this API is still marked as unstable,
developers should be reluctant about adopting this API in the first
place.)