Fixed Whitesource configuration
Some of the error messages observed when running Whitesource were caused by an incorrect configuration. The configuration is now correct, and the remaining errors have to be manually sorted using the Whitesource web interface. The dev dependencies are not curently included, and local resolution (i.e. without querying NPM) is enforced.