Fix getThingAll() not grouping Quads by Thing
When a LitDataset includes multiple Quads with the same Subject,
those Quads together make up a Thing. However, because NamedNode
objects with the same IRI are actually different instances, a Set
is unable to deduplicate them - where it _can_ do so with strings.
This resulted in getThingAll() actually returning a separate Thing
for every Quad, rather than for every distinct Subject Node.

This commit fixes this by looping over all seen Nodes, and only
adding new Nodes if the list of seen Nodes does not include it yet.