Use consistent marker of inaccessible ACLs
A LitDataset can have an `acl` object attached with the keys
`resourceAcl` and `fallbackAcl`, which are undefined and null
respectively if they are not found, the cause of which is usually
lack of Control access for the current user.

That will also result in the original Resource not even advertising
an ACL. In that case, the `acl` object as a whole was set to
`null`, meaning that the result could no longer be passed to
getAgentAccessModesOne/All without casting, even though that is
perfectly capable of dealing with inaccessible ACLs (it will just
return null).

So now, if we can deduce from a Resource's metadata that neither
ACL file is going to be accessible anyway, we mark that in the same
way as we would mark it if we had tried to fetch those files and
_then_ found out that they were inaccessible.