Limit Notification sugar to single Things
Supporting building entire LitDatasets became rather messy, and
hardly easier than just taking the generated Thing, adding it to
a LitDataset yourself and saving that to an Inbox Container.
Given that a Notification consisting of multiple Things is likely
to be a more advanced use case (most examples in the spec do not do
it), I think it's fine to have the developer deal with that
manually.