Back out notification code
The way we currently implement Notifications is not in line with
the spec. Specifically, we set the Target to the Resource that is
pointing to the Inbox, and suggest that the Origin is the current
user's WebID. However, if the Notification is e.g. about a Move
Action, they should be the source and target of the Move.

For more information, see:
https://www.w3.org/TR/activitystreams-vocabulary/#origin-target

Thus, we decided to back out these changes for now, in lieu of
implementing it properly.