Mark internal properties as internal_
We attach some properties to LitDataset and Thing that we do not
intend for the developer to access directly, because we will likely
be moving that data into the Dataset itself in the future. Instead,
we provide functions to access that data.

To more explicitly discourage direct access to those properties,
they are now tagged as internal as well.