Fix access granting bug when ACL is both default and resource (#290)
If an ACL is both a default and a resource rule, an additional check must be run to verify that setting resource access for the agent doesn't escalate his default access.

The exact same bug would have happened when setting default access to an agent from a rule being both a resource and a default one, so the same logic needed to apply. In order to avoid duplication, the code snippet is isolated in its own function and applied twice.