Align rule-splitters for Agents and the Public
This is somewhat complex code that is mostly equivalent to both,
and will also be mostly equivalent for when we write support for
setting Group access and other potential future targets. Thus, to
prevent confusion, make sure that the functions actually _look_
mostly the same in addition to _behaving_ mostly the same.