More helpful error on incorrect saveAclFor() calls
Developers not using TypeScript will not be warned when calling
saveAclFor() withh a Resource that does not expose an ACL URL to
the current user, so they'll have to try it out and see it fail in
the right circumstances. This should make the error message that
then occurs provide more assistance as to how to resolve it.