Return null when new Resource is not accessible
This better represents the reality of not being able to access the
newly-created Resource, in which case you can only access the data
you've actually sent. The downside of this is that you will no
longer be able to access the URL of the newly-created Resource if
you're not able to Read from it, but that seems OK given that
there's not much to do with it.