Remove support for Constant Policies
No server is implementing yet, and neither is the name finalised.
Since other code would build on top of this (and would thus also
have to do rework later), I'm removing this for now. We can revert
this commit and implement support in higher layers when we are
actually supporting them later.