Distinguish between sent and received resourceInfo
This adds a type WithServerResourceInfo which is like the existing
WithResourceInfo, but adds properties that the server has to tell
us. Specifically, it is data received when sending a request to
that specific Resource (like the URLs of linked Resources), but not
when creating the Resource by sending a POST to its Container.

Since all required properties that were available on
WithResourceInfo in the last release are still available now, this
allows us to make the previous breaking change no longer breaking.
Instead, we explicitly clarify that the return value of our
`save*()` functions are the data we _sent_ to the Pod, which is not
necessarily all data that's actually available at the Pod (e.g.
when someone changed it in between our fetching and saving it).

When functions _can_ obtain data that's set by the server (i.e.
when doing a request directly to the Resource), they now return
WithServerResourceInfo, which is an extension of the existing
WithResourceInfo that (currently) adds a list of Linked Resources.