Make Access Control-functions work on WithAcrs
If Solid ever gets updated to be able to send both a Resource and
its Linked Resources in a single Response, we'll have API
compatibility if we treat a Resource and its Linked Resources as a
single object.

We've done that for ACLs so far, but I forgot to treat ACRs in the
same way. This commit fixes that: operations on ACRs now take a
Resource with an attached ACR as parameter.