No longer automatically fetch attached Policies
Since they're not server-managed and will therefore never be
returned by the server together with the actual data in a single
HTTP Response, the Access Policy Resources should not be
automatically fetched by us.

Instead, we provide a convenience function that will give the
developer a list of all the existing Policy Resources that are
being referenced in an ACR, so they can fetch them themselves (and
potentially deal with errors that come up when they aren't
accessible to the current user).