Allow preview access to ACP functions
Although ACP is not ready for production use in the sense that the
API will still be changing without warning, it is available for
preview testing on the acp_lowlevel_preview object. Note that this
does not facilitate tree-shaking: if you use one of the functions,
all of them will be included in your bundle.