End-to-end test for Container creation in ESS
The bug that initially caused this test to be disabled is resolved,
and the new end-to-end test verifies this. However, since
we cannot do authenticated requests in Node yet, and since there is
no WAC-powered ESS instance available yet that allows us to write
data, that end-to-end test is part of the browser-based end-to-end
test suite.