Workaround for N3's import
This is required to make it work in Node with ES module support,
where it only has a default export (because it is published as a
CommonJS module there), and in Webpack, where it only has named
exports (because it is published as an ES module there).

Since dynamic imports always return a single object (with the
named exports as properties for ES modules, and with a `default`
property containing all exports for CommonJS modules), we can use
that to import the module and then dynamically determine whether
we have to return the `default` property or the imported object
itself.

Unfortunately, that does mean that tree-shaking N3 does not work,
until N3 makes itself compatible with both Node's module systems
and the ES imports from Webpack and co:
https://github.com/rdfjs/N3.js/issues/196