Disable end-to-end tests against ESS
Since ESS switched to ACPs we no longer have a convenient way to
prepare the tests data with the proper permissions (i.e.
public-read-write). Therefore, end-to-end tests against ESS have
been disabled for now.  We can re-enable them once we have a Node
library with which we can authenticate, after which we can set the
relevant permissions in the tests themselves.