Run end-to-end tests against the published package
Although we're not running the tests as thoroughly as we do for
in regular CI (i.e. in different Node versions and OS's), running
them once with code from the published package should be enough to
help us detect issues that arose in the build process, such as the
output directory not matching the files listed in the package.json,
or type definitions that are missing.