Make end-to-end tests work with Cognito signin
Note that I also had to explicitly pass in window.fetch for the
unauthenticated request; without it, I'd encounter the following
error:

> Invalid character in header field name

Apparently, some polyfill has trouble with headers on a 4xx
response. Since I wasn't able to reproduce that outside of the test
fixture, I didn't investigate that further.