Export ACL-related API's from ACL module
With the export map now making the module in which an API is
defined relevant to the consumer, it makes sense to keep all our
ACL-related API's in the ACL module.

The only remnant is some references to ACLs in the Resource
metadata, but that can be removed once the ACL-related API's have
been updated to look at the generic Linked Resources metadata.