Make CodeSandboxes work again
Before this, running CodeSandboxes directly from the repo (i.e. not
from a link in a pull request, in which CodeSandbox has replaced
the dependency on @inrupt/solid-client with a version that it built
itself, based on the code in the PR) resulted in the following
error:

> Could not fetch dependencies, please try again in a couple
> seconds: Could not fetch
> https://cdn.jsdelivr.net/gh/../..//package.json

The reference to the local package was only there for the TestCafe
tests, so we now install it there explicitly.