Migrate ACP end-to-end tests to Node
Now that we can send authenticated requests in Node, we can
manipulate Access Control Policies there. Since browser-based tests
are relatively brittle [1] and pretty involved (the interaction
between the test script and the in-browser code, i.e. the
ClientFunctions), we limit those to a small smoke test that
verifies that we can read and manipulate data while logged in.

[1] https://testing.googleblog.com/2017/04/where-do-our-flaky-tests-come-from.html