Commit optionality of dependencies to lockfile
Weirdly enough, after dependabot updates, if I ran npm install in
the root directory the package-lock.json would be updated with a
couple of `"optional": true` properties for a number of
dependencies, even though `npm ci` ran successfully.

I am not 100% what the cause of this would be. My first theory was
that dependabot used a different version of npm, but since
installing Node 14 with NVM doesn't seem to update the npm version,
it seems unlikely that dependabot was using v7 already.

My current hypothesis is that this was an instance of this bug:
https://github.com/npm/npm/issues/17722
The tl;dr, I believe, is that a package that was not optional in
dependabot's environment is optional in mine, and npm then keeps
updating the lockfile. Unfortunately, there does not seem to be a
resolution yet.