Fix updating of Things containing Blank Nodes
This bug was caused by a combination of two factors:

1. When you fetch a SolidDataset twice, the two instances of the
   same Blank Node inside those two different SolidDatasets are not
   considered equal.
2. When a developer calls setThing(), we first delete all that
   Thing's existing Quads from the SolidDataset, and then re-add
   the ones of the given Thing. To prevent the same Quad from being
   both deleted and added in the same request, after re-adding a
   Quad we check if that Quad was already listed in the deletions
   of the ChangeLog, and if it is, rather than adding it to the
   ChangeLog's additions, we remove it from the deletions.

Now, checking if a Quad was listed in the deletions would fail for
Quads containing Blank Nodes if this was a different instance of
the same SolidDataset, because those Blank Nodes are considered
different.

To get around this, I've changed the code such that, when looking
whether a Quad was already listed in the deletions, I now consider
any two Blank Node objects equivalent. This should be safe, because
when a Quad with a Blank Node is listed in the additions, that can
only happen because it was already present in the first place.
After all, solid-client does not support adding Blank Nodes. (Well,
technically you can manually pass it one, but since that Blank Node
is not used as a Subject anywhere, that would not do anything.)