Add getAgentAccess for WAC resources
This adds the getAgentAccess function to list permissions an Agent is
given for a Resource. This function fecthes the metadata, so it deals
with the notion of fallback ACL internally. Note that in order to
accomodate for the universal API, this adapts the notion of Access as
initially defined in the ACL logic to align with the fact that in ACP, an
undefined access isn't granted, while in ACL, a false value was used.

* Create explicit type for access from WAC: WacAccess is a type based on the universal Access (i.e. ACP Access), and
adds WAC-specific restrictions.

Co-authored-by: Pat McBennett <patm@inrupt.com>