getActorAccess only deals with explicit access
Since we cannot reliably calculate *effective* access for a given
actor (for example, we cannot know whether an Agent is also in a
particular Group), getActorAccess() only concerns itself with
access that has *explicitly* been defined for the given actor.

However, this was not applied consistently. If additional
restrictions were added through an allOf Rule (for example, when
checking access for a particular Group, if you had one allOf Rule
determining that you had to be in that Group, but another one that
said you also had to be in another Group), then getActorAccess()
would consider the Policy not to apply. However, if additional
restrictions were applied through noneOf Rules, then those Rules
would simply be disregarded unless they mentioned the given actor.
What this meant was that, if you were e.g. checking the accesss
given to a particular group, it would say that the access defined
in that Policy applied _even if it might not_, for example when
the Agent was listed in a noneOf Rule.

This inconsistency is now corrected: if further restrictions
prevent a Policy from *always* applying to the given actor, then
getActorAccess() no longer considers it. (And likewise for
getActorAccessAll().)