Add getAgentAccessAll for WAC Resources
This adds `getAgentAccessAll`, which returns a list of all the
individual agents being granted access to a Resource which access is
controlled by a WAC server.

Note that since the WAC code is largely based on pre-existing functions,
the tests check for implementation rather than behaviour by verifying
that the underlying function is called, in order to reduce the number of
tests.