Keep exhaustive list of actorRelations
In several place in the code, we will be copying over existing
Rules excluding a specific actor. This requires iterating over all
actorRelations and copying over those that are *not* the given
actor. However, since the ACP spec might be expanded with
additional actorRelations, it is important that we update our code
in all the relevant places when we support a new one.
Therefore, this introduces a single place to keep track of all
of them, and refers to that everywhere we do something with
actorRelations.
This ensures that when, in the future, when we intend to add
support for an additional actorRelation, we will encounter this
type, add the desired relation there, and automatically have all
code either ready to go, or at least be pointed out by TypeScript
warnings to indicate they need updating.