Add getAccessFor, wrapper for libraries
This adds `getAccessFor`, a wrapper around `getAgentAccess`,
`getGroupAccess` and `getPublicAccess` that is convenient to call from a
library, because it's only one function, and the appropriate underlying
function is determined based on the parameters.

It is implemented in a separate file from the functions it calls to
mainly to make it easier to test: the whole "universal" module can be
mocked. If `getAccessFor` is implemented alongside e.g.
`getAgentAccess`, jest cannot mock the implementation of the latter when
testing the former.