Fix removal of ACP Policies setting desired access
This bug was found by fast-check: when a Policy already existed
that defined the access we wanted to set, that Policy would be
removed. We now just leave this Policy intact.

We did already add new Policies with the desired access, but only
did so if the desired access actually conflicted with the existing
access. This is desired: we want to re-use existing Policies if
possible, rather than removing them and recreating new ones that do
the same.

The other two recorded counterexamples were instances of the same
bug, so I removed those. fast-check did find a new counterexample,
so I've added that as a (currently skipped) test case to be fixed
later.