Fix removal of access without replacing it
Another bug discovered by fast-check.

If:
- An access mode was requested to be set to false,
- A Policy exists that denies that access,
- Another Policy exists that allows that access, also applies to
  another actor, and allows other desired access

then that latter Policy would be unapplied (because it allowed
access we did not want to apply), but no new Rule would be added to
apply the other desired access (because the existing combination of
Rules already resulted in that access).

Now, we first simulate removal of the access, then check whether
the desired access would still apply, and if not, then we *do* add
a new Rule applying the desired access.