Prevent overriding resource if IRI ends with #
When fetching a resource, its source IRI is server-managed, which means
it does not include a hash fragment. Writing back to a resource that
has been fetched including the hash fragment, there was a mismatch
between the source IRI (no fragment) and the target IRI (with the
fragment), which lead the library to do a PUT instead of a PATCH.

IRIs are now normalized before being compared.