Disregard user-supplied IRI to set resource info
The source IRI of a give resource is a piece of server-managed
information, and as such it should not be derived from the user-provided
parameter to the called function. When a resource has been created, the
response to the creation request contains a target URL, against which
one can resolve the IRIs in the response headers if they are relative.

This involves some refactoring in the tests, where the resource does not
include such target URL by default.