Prevent overriding resource if IRI ends with # (#942)
When fetching a resource, its source IRI is server-managed, which means
it does not include a hash fragment. Writing back to a resource that
has been fetched including the hash fragment, there was a mismatch
between the source IRI (no fragment) and the target IRI (with the
fragment), which lead the library to do a PUT instead of a PATCH.
IRIs are now normalized before being compared.

- Disregard user-supplied IRI to set resource info: The source IRI of a give resource is a piece of server-managed
information, and as such it should not be derived from the user-provided
parameter to the called function. When a resource has been created, the
response to the creation request contains a target URL, against which
one can resolve the IRIs in the response headers if they are relative.
This involves some refactoring in the tests, where the resource does not
include such target URL by default.

- Apply normalization function to ACP: This also enforces that ACP use the target URL too

- Test that response URL is used: This adds explicit tests verifying that the response URL is used over
the user-provided target URL