Only run secret-requiring jobs when available
Specifically, Dependabot does not have access to our repository
secrets (because they might leak to malicious dependencies), so
it cannot publish to npm or run tests that require a Pod's
credentials. Thus, if a PR comes in that requires verification that
e.g. publication is not broken, someone with sufficient access will
have to trigger the jobs manually.

See
https://github.blog/changelog/2021-02-19-github-actions-workflows-triggered-by-dependabot-prs-will-run-with-read-only-permissions/