Fix stricter TestCafe's ClientFunction confusion
We have no special configuration for TestCafe to use TypeScript, so
presumably it is using its own configuration, which appears to have
gotten stricter from one day to another. Either way, referring to
variables in a ClientFunction that is not in scope of the parent
script started throwing an error; this fix creates a variable of
the same name in the parent script.