Only skip prepare-deployment steps for Dependabot
Before, I marked the job as a whole as being skippable for
Dependabot, because it is not required for Pull Requests. However,
since other jobs that *are* required depend on it, skipping it
would also skip those jobs.

Thus, now only a step in the prepare-deployment job is skipped when
the actor is dependabot, so that the job as a whole still appears
to succeed.