Only nest Blank Nodes used for a single Subject
Like in Turtle, the replacing of Blank Nodes with combinations of
Predicates and Objects only really works for single Subjects - if
multiple Subjects want to reuse the same Blank Nodes, that Blank
Node will need to have its own identifier.

Thus, rather than having an exception for specific types of Blank
Nodes to *not* use those in nested structures, this commit turns
that upside down: we only specify which Blank Nodes we *do* use
in those structures, i.e. those that only appear for a single
Subject, and are not in a cycle.