Add proptests for RDF/JS Datasets with Blank Nodes
Unfortunately, Blank Nodes being annoying that way, it's not
possible to directly compare RDF/JS Datasets before and after
serialisation, because Blank Nodes will then have a different
identity.

However, we can specify some general properties that should hold:
- Before and after serialisation, the Dataset should have the same
  number of Quads.
- Everything that is not a Blank Node should be present both before
  and after serialisation.

Thus, this adds a proptest that checks for that. One thing to note
is that we do need to make sure that it's possible that not every
Blank Node in a generated Dataset is unique, but the current way
we do that is not foolproof: we generate just a single Blank Node
and replace arbitrary Blank Nodes with that one with an arbitrary
probability. This means that e.g. cycles can not appear.

A possible other approach to consider is to just generate a number
of Blank Nodes for a Dataset, and then replace some arbitrary
Subjects and Objects with some of those - though there's probably
caveats with that as well.