Temporarily disable end-to-end tests against ESS
Inrupt's Enterprise Solid Server changed its authentication
mechanism to invalidate the refresh token after every use,
returning a new one to use instead. This does not mesh well with
our CI setup, which used a token set up once in the CI's
environment secrets.

The server team will probably be able to arrange a new setup for us
that does not require maintaining state across test runs, so while
we await that, we'll disable automated tests against it for a short
while.