Make SolidDatasets and Things atomic objects
This allowed for the removal of many brittle heuristics, allows
for better integration with the wider ecosystem, ensures that no
methods are present but intended not to be used, makes debugging
easier for developers, and is likely to improve performance (though
this wasn't measures). It also allows developers to obtain the data
in a way that integrates with the wider RDF/JS ecosystem.