Produce valid URLs when creating Policies/Rules
When we generate new Policies or Rules, we based their URLs on the
URL of the actor to which it applies. However, we did not URL
encode it in every case, so if those URLs included a `#`, like
WebIDs often do, that would result in invalid URLs with multiple
hash fragments.