Loosen the type of saveAclFor
It was unnecessarily strict - any SolidDataset can be used as an
ACL, not just ones that were used as an ACL already.

Specifically, this allows setting the ACL for a Resource that did
not have one yet (to a SolidDataset newly created with
createSolidDataset()), rather than merely updating the existing
one.