Allow experimenting with responseToSolidDataset
Although we haven't publicly documented it yet (and hence it does
not come with our usual stability guarantees - i.e. we can change
the API in non-major versions), by exporting it people can start
experimenting with it to give us feedback on whether it's useful
and how it can be improved.