Upgrade to Parcel v2
It's technically not marked as stable yet, but since its developers
are not really supporting v1 anymore and tell everybody to use v2,
it might as well be.

Case in point: https://github.com/parcel-bundler/parcel/issues/5943#issuecomment-789080294
Also the warning messages Parcel now displays in the CI logs
recommending an upgrade to v2.