Fix missing build-time error for asUrl()
With ThingLocal now being a subset of ThingPersisted thanks to its
use of a skolemised URL, editors could no longer recognise that a
ThingLocal passed to asUrl() also required a baseUrl. With this
updated type definition, developers will once again have
red squigglies if they forget to pass a baseUrl together with a
ThingLocal.

Unfortunately, a Thing of which it is unknown whether it's a
ThingLocal or a ThingPersisted is indistinguishable from a
ThingPersisted, so in those cases developers will still have to
rely on the runtime error.