Fix missing build-time errors for asUrl()
With ThingLocal now being a subset of ThingPersisted thanks to its
use of a skolemised URL, editors could no longer recognise that a
ThingLocal passed to asUrl() also required a baseUrl. With this
somewhat involved type definition, developers will once again have
red squigglies if they forget to pass that.