Resolve npm audit resolutions once and for all
This automatically runs npm-force-resolutions before installing
something, to hopefully keep transitive dependencies at a
non-vulnerable version even if their parent dependencies have not
updated yet, even across multiple dependency updates.

Unfortunately, since it needs to run before installing something,
we can't add it as a direct dependency, so I had to run it using
`npx` and pin the version in there.

I'm not sure yet if a preinstall script will work for maintaining
it across dependabot updates and with `npm ci` runs, but I figured
we can at least try and see if it works.