Only prepare end-to-end tests when running them
This prepare job fails relatively often when npm is trying to
install things and fails fetching the relevant metadata. Since
we're not actually running the browser-based end-to-end tests all
the time, we should be able to cut down on those failures by only
running the preparation step when we're actually going to run
browser-based end-to-end tests.