Manually fetch all ACP Data
Rather than just bailing out whenever an Access Control Resource
references Policies and/or (indirectly) Rules located outside of
the ACR, we now attempt to fetch them all. It is up to whoever
creates those Policies and Rules to avoid us having to make lots of
requests to determine access.

The way I implemented this is by adding a new function that reads
all references to Policies in an Access Control Resource, fetches
those, then reads all references to Rules in those Policies and
fetches _those_, and then collects them all in a single data
structure that contains all the data that is needed to determine
access to a given Resource. This allows the existing (complex)
functions to determine access to stay mostly the same, and limits
hairy asynchronous code to a single place (getPoliciesAndRules).

The API will still bail out if one or more of the Policies or Rules
could not be retrieved, e.g. because the current user does not have
access to it.