Fix base case for cycle detection
When multiple Blank Nodes referenced each other, they would
erroneously be marked as being part of a cycle and hence not a
chain, leading to the chain detection algorithm not working and the
blank nodes being listed at the top level of the graph with a
non-deterministic identifier. This fixes that; if those Blank Nodes
for a chain without cycles, they will now be represented as a tree
without needing to generate identifiers for them.