Account for Link with rel="acl" pointing to ACRs
The ACP proposal is going to be updated such that Link headers with
rel="acl" can also point to Access Control Resources, rather than
being used by WAC exclusively in order to point to ACLs.

I'm still not entirely sure that we've covered all potential
problems this may cause with this change, but I expect it to be
enough.

Ideally, WAC would also mandate a rel="type" Link header for Access
Control Lists, so we could not just check that it is _not_ of type
ACR, but explicitly check that it _is_ of type ACL.