Policies only match with anyOf or allOf Matchers
Previously, noneOf Rules would result in a match even if there were
no allOf and anyOf Rules. It turns out we didn't implement that
properly, but now, the intended behaviour is that Matchers *only*
make a Policy apply if there is at least a match based on an
allOf or anyOf Matcher.