Fix mocks in test
This ensures that the mocks are properly configured when testing to
write a key to a profile document. In particular, some functions of the
resource module are stubbed as jest.fn(), and each test using them may
add its desired behaviour on this stub. Note that the appropriate
behaviour is also observed based on the mocks: since this function does
not return, but rather it saves the updated profile to the Pod, we read
the updated value from the mock save.