getThingAll returns blank nodes too (#1273)
getThingAll used to only return Things which subject is a named node. This is an issue in the case where one legitimately wants to get values from a blank node at the "top level" of a graph, e.g. when parsing the .well-known/solid document, which only contains one blank node. The blank node identifier being by nature unpredictable, getThingAll is the only way to access such blank node.