Export RDF parser functions (#1318)
This exports the RDF parsers that may be passed to `getSolidDataset`. In particular, this is useful to fetch non-Solid resources that are not available in Turtle, which is the default format. This is in particular the case of some .well-known discovery documents, such as .well-known/vc-configuration, which is only available in JSON-LD.