Discover personal profile from WebID profile doc
This adds two functions to the API which enable discovering personal profile documents from WebID profile documents. The purpose is to make it easier to link from a non-Solid WebID profile document to a Solid personal profile, so that the WebID profile may not be editable but the personal profile is.