Discover personal profile from WebID profile doc (#1335)
This adds two functions to the API which enable discovering personal profile documents from WebID profile documents. The purpose is to make it easier to link from a non-Solid WebID profile document to a Solid personal profile, so that the WebID profile may not be editable but the personal profile is.

The output contains both the WebID profile document and the potential alternative profiles if any have been found. This prevents from having to fetch them separately, and it also clarifies what data is available where.

Co-authored-by: Matthieu Bosquet <matthieu.bosquet@inrupt.com>