Enable e2e test public access (#1375)
This re-enables E2E ACP testsusing environment-scoped secrets, instead of global repository ones. 

Note that depending on the ESS version being run in each environment, certain version of our APIs should be used, and that even with this a test seems to fail. Will investigate and fix in a different PR.

Co-authored-by: Nicolas Seydoux <nseydoux@inrupt.com>