Remove unnecessary type assertion in getProfileAll
This introduces an overload to cleanly define the return type of getProfileAll, namely it the WebID profile is NOT provided, the return type includes that of `getSolidDataset`, and if it IS provided then the return type includes the generics instanciated by the provided WebID profile. Making this distinction allows to remove a type assertion from the implementation.